Moved test classes around, renamed, merged etc to reflect tested project structure
(cherry picked from commit c51929b4a5b94592b1ceafad4cf865c3215b7bcd)

Conflicts:
	src/WebServer.UnitTests/Http/HttpServerTests.HandleRequest.cs
