+++ src/HttpMessage.UnitTests/HttpMessage.UnitTests.csproj
@@ -95,10 +95,10 @@
     <SDKReference Include="TestPlatform.Universal, Version=$(UnitTestPlatformVersion)" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="HttpServerRequestTest.cs" />
-    <Compile Include="HttpServerResponseTest.cs" />
+    <Compile Include="HttpServerRequestTests.cs" />
+    <Compile Include="HttpServerResponseTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestHelpers\TestInputStream.cs" />
+    <Compile Include="_TestHelpers\TestInputStream.cs" />
     <Compile Include="UnitTestApp.xaml.cs">
       <DependentUpon>UnitTestApp.xaml</DependentUpon>
     </Compile>
