+++ src/WebServer.UnitTests/File/MimeTypeProviderTests.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
 using Restup.Webserver.File;
+using Restup.Webserver.UnitTests.TestHelpers;
 using System;
 using System.Collections.Generic;
