+++ src/WebServer.UnitTests/Rest/RestRouteHandlerTest.Happy.cs
@@ -6,7 +6,7 @@ using System;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Restup.Webserver.UnitTests
+namespace Restup.Webserver.UnitTests.Rest
 {
     [TestClass]
     public class RestRouteHandlerHappyPathTest
