+++ src/WebServer.UnitTests/TestHelpers/HttpServerTests.Fluent.cs
@@ -2,32 +2,31 @@ using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
 using Restup.HttpMessage;
 using Restup.HttpMessage.Headers.Response;
 using Restup.Webserver.Http;
-using Restup.Webserver.UnitTests.TestHelpers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Restup.Webserver.UnitTests.Http
+namespace Restup.Webserver.UnitTests.TestHelpers
 {
-    public class HttpServerFluentTests
+    public class HttpServerTests_Fluent
     {
         private readonly HttpServer _httpServer;
         private readonly List<HttpServerResponse> _responses;
 
         private readonly Dictionary<string, EchoRouteHandler> _routeHandlers;
 
-        public HttpServerFluentTests Given => this;
-        public HttpServerFluentTests When => this;
-        public HttpServerFluentTests Then => this;
+        public HttpServerTests_Fluent Given => this;
+        public HttpServerTests_Fluent When => this;
+        public HttpServerTests_Fluent Then => this;
 
-        public HttpServerFluentTests()
+        public HttpServerTests_Fluent()
         {
             _httpServer = new HttpServer(80);
             _responses = new List<HttpServerResponse>();
             _routeHandlers = new Dictionary<string, EchoRouteHandler>();
         }
 
-        public HttpServerFluentTests ListeningOnDefaultRoute()
+        public HttpServerTests_Fluent ListeningOnDefaultRoute()
         {
             var routeHandler = new EchoRouteHandler();
             _routeHandlers[string.Empty] = routeHandler;
@@ -35,7 +34,7 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests ListeningOnRoute(string urlPrefix)
+        public HttpServerTests_Fluent ListeningOnRoute(string urlPrefix)
         {
             var routeHandler = new EchoRouteHandler();
             _routeHandlers[urlPrefix] = routeHandler;
@@ -43,7 +42,7 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests RequestHasArrived(string uri, IEnumerable<string> acceptEncodings = null, byte[] content = null)
+        public HttpServerTests_Fluent RequestHasArrived(string uri, IEnumerable<string> acceptEncodings = null, byte[] content = null)
         {
             var httpServerRequest = Utils.CreateHttpRequest(uri: new Uri(uri, UriKind.Relative), acceptEncodings: acceptEncodings, content: content);
             var response = _httpServer.HandleRequestAsync(httpServerRequest).Result;
@@ -51,7 +50,7 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests AssertLastResponse<T>(Func<HttpServerResponse, T> actualFunc, T expected)
+        public HttpServerTests_Fluent AssertLastResponse<T>(Func<HttpServerResponse, T> actualFunc, T expected)
         {
             var response = _responses.Last();
             var actual = actualFunc(response);
@@ -60,7 +59,7 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests AssertLastResponse<T>(Func<ContentEncodingHeader, T> actualFunc, T expected)
+        public HttpServerTests_Fluent AssertLastResponse<T>(Func<ContentEncodingHeader, T> actualFunc, T expected)
         {
             var response = _responses.Last();
             var header = response.Headers.OfType<ContentEncodingHeader>().First();
@@ -70,7 +69,7 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerRequest<T>(Func<IHttpServerRequest, T> actualFunc, T expected)
+        public HttpServerTests_Fluent AssertRouteHandlerRequest<T>(Func<IHttpServerRequest, T> actualFunc, T expected)
         {
             var routeHandler = _routeHandlers.Single().Value;
             var request = routeHandler.Requests.Last();
@@ -80,21 +79,21 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerReceivedRequest()
+        public HttpServerTests_Fluent AssertRouteHandlerReceivedRequest()
         {
             var routeHandler = _routeHandlers.Single().Value;
             Assert.AreEqual(1, routeHandler.Requests.Count());
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerReceivedNoRequests()
+        public HttpServerTests_Fluent AssertRouteHandlerReceivedNoRequests()
         {
             var routeHandler = _routeHandlers.Single().Value;
             Assert.AreEqual(0, routeHandler.Requests.Count());
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerRequest<T>(string prefix, Func<IHttpServerRequest, T> actualFunc, T expected)
+        public HttpServerTests_Fluent AssertRouteHandlerRequest<T>(string prefix, Func<IHttpServerRequest, T> actualFunc, T expected)
         {
             var routeHandler = _routeHandlers[prefix];
             var request = routeHandler.Requests.Last();
@@ -104,14 +103,14 @@ namespace Restup.Webserver.UnitTests.Http
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerReceivedRequest(string prefix)
+        public HttpServerTests_Fluent AssertRouteHandlerReceivedRequest(string prefix)
         {
             var routeHandler = _routeHandlers[prefix];
             Assert.AreEqual(1, routeHandler.Requests.Count());
             return this;
         }
 
-        public HttpServerFluentTests AssertRouteHandlerReceivedNoRequests(string prefix)
+        public HttpServerTests_Fluent AssertRouteHandlerReceivedNoRequests(string prefix)
         {
             var routeHandler = _routeHandlers[prefix];
             Assert.AreEqual(0, routeHandler.Requests.Count());
